refactor hwmd project to submodule bumps to 1.0.0-alpha
