+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Calendar Versioning](https://calver.org/#scheme) scheme (`YYYY.MM.PATCH`).
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## prod
 [2022.12.2-beta]
@@ -12,6 +12,12 @@ and this project adheres to [Calendar Versioning](https://calver.org/#scheme) sc
 
 ----
 ## [Unreleased]
+
+## [1.0.0-alpha]
+- [change] calver to semantic versioning
+- [change] python filenames
+- [remove] build shell script
+- [remove] Makefile build and boot commands
 
 ## [2022.12.2-beta]
 _(dmidecode = 3.3 | smartmontools = 7.2 | hwinfo = 21.72 | lshw = 02.19 | lspci = 3.7.0)_
