+++ hwmd.py
@@ -1,9 +1,7 @@
 import subprocess
 import json
 
-
-
-class HWMDRetrieval:
+class HWMD:
     """ Collect as much hardware data as possible using lshw, dmidecode, lspci and others tools."""
 
     def get_lshw_data(log):
